+++ editor/editor-plugins/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-plugins
 
+## 13.0.78
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 13.0.77
 
 ### Patch Changes